Make the Stream type hold a Ptr rather than ForeignPtr
This is a quick workaround for the issue where the finalizer
(lzma_end) can run prematurely causing random segfaults. This patch
fixes it by moving the responsibility of running the finalizer
from the RTS to the user. The user is expected to run the finalizer
at the end of Stream's life cycle. Otherwise the program leaks memory.

This issue should really be fixed by managing its life cycle properly
while keeping the Stream type hold a ForeignPtr.